Alter field quantity_in_stock on product

+++ ecommerce_api/products/models.py
@@ -5,5 +5,5 @@ class Product(models.Model):
     name = models.CharField(max_length=255)
     description = models.TextField()
     price = models.DecimalField(max_digits=10, decimal_places=2)
-    quantity_in_stock = models.FloatField()
+    quantity_in_stock = models.IntegerField()
     image = models.ImageField(upload_to='product_images/', null=True, blank=True)
